+++ module-apps/windows/AppWindow.cpp
@@ -11,7 +11,7 @@ using namespace style::header;
 namespace gui
 {
 
-    AppWindow::AppWindow(app::Application *app, std::string name, uint32_t id) : Window(name, id), application{app}
+    AppWindow::AppWindow(app::Application *app, std::string name) : Window(name), application{app}
     {
         setSize(style::window_width, style::window_height);
     }