+++ module-db/Interface/SMSRecord.cpp
@@ -2,6 +2,7 @@
 #include "Common/Query.hpp"
 #include "ContactRecord.hpp"
 #include "ThreadRecord.hpp"
+#include "queries/sms/QuerySMSAdd.hpp"
 #include "queries/sms/QuerySMSGet.hpp"
 #include "queries/sms/QuerySMSGetByContactID.hpp"
 #include "queries/sms/QuerySMSGetByID.hpp"
@@ -9,6 +10,7 @@
 #include "queries/sms/QuerySMSGetCount.hpp"
 #include "queries/sms/QuerySMSSearch.hpp"
 #include "queries/sms/QuerySMSRemove.hpp"
+#include "queries/sms/QuerySMSUpdate.hpp"
 #include "queries/sms/QuerySMSSearchByType.hpp"
 #include <log/log.hpp>
 
@@ -317,8 +319,14 @@ std::unique_ptr<db::QueryResult> SMSRecordInterface::runQuery(std::shared_ptr<db
     else if (typeid(*query) == typeid(db::query::SMSGetCount)) {
         return getCountQuery(query);
     }
+    else if (typeid(*query) == typeid(db::query::SMSAdd)) {
+        return addQuery(query);
+    }
     else if (typeid(*query) == typeid(db::query::SMSRemove)) {
         return removeQuery(query);
+    }
+    else if (typeid(*query) == typeid(db::query::SMSUpdate)) {
+        return updateQuery(query);
     }
     else if (typeid(*query) == typeid(db::query::SMSGet)) {
         return getQuery(query);
@@ -414,11 +422,31 @@ std::unique_ptr<db::QueryResult> SMSRecordInterface::getCountQuery(std::shared_p
     response->setRequestQuery(query);
     return response;
 }
+std::unique_ptr<db::QueryResult> SMSRecordInterface::addQuery(std::shared_ptr<db::Query> query)
+{
+    const auto localQuery = static_cast<const db::query::SMSAdd *>(query.get());
+    auto record           = localQuery->record;
+    const auto ret        = Add(record);
+    if (ret) {
+        record.ID = GetLastID();
+    }
+    auto response = std::make_unique<db::query::SMSAddResult>(record, ret);
+    response->setRequestQuery(query);
+    return response;
+}
 std::unique_ptr<db::QueryResult> SMSRecordInterface::removeQuery(std::shared_ptr<db::Query> query)
 {
     const auto localQuery = static_cast<const db::query::SMSRemove *>(query.get());
-    auto ret              = smsDB->sms.removeById(localQuery->id);
+    auto ret              = RemoveByID(localQuery->id);
     auto response         = std::make_unique<db::query::SMSRemoveResult>(ret);
+    response->setRequestQuery(query);
+    return response;
+}
+std::unique_ptr<db::QueryResult> SMSRecordInterface::updateQuery(std::shared_ptr<db::Query> query)
+{
+    const auto localQuery = static_cast<const db::query::SMSUpdate *>(query.get());
+    auto ret              = Update(localQuery->record);
+    auto response         = std::make_unique<db::query::SMSUpdateResult>(ret);
     response->setRequestQuery(query);
     return response;
 }