+++ module-utils/utf8/UTF8.hpp
@@ -98,6 +98,10 @@ class UTF8
     {
         return strLength;
     }
+    bool empty() const noexcept
+    {
+        return strLength == 0U;
+    }
     uint32_t used() const
     {
         return sizeUsed;